Add asset name and issuer to labels
This change adds the issuer account and asset name to the prometheus label `asset_name` in the format `<issuer_account>:<asset_name>` for assets that are not native.  For native assets, the `asset_name` is set to "XLM" to ensure that an asset_name label can be relied upon in all instances.